Finish some system files; label improvements; add text and dialogue bytecode extraction

+++ src/game/setCutscenes.h
@@ -155,7 +155,7 @@ extern u16 func_800A7AE8(u8);
 extern u16 func_800A7DFC();        
 extern u16 func_800A87C4();   
 extern void func_800A8F74();   
-extern void loadCutscene(u32);  
+extern void loadCutscene(u8);  
 
 // maybe game.c bss
 extern u16 gCutsceneIndex;
